requirement.txt
Django>=4.2,<5.0
djangorestframework
djongo
pymongo
